fix small bug on jit log dump code
Former-commit-id: e06a4e920898fd2532d2d9683593d1bb2152b288

+++ src/jit/ESIR.h
@@ -299,7 +299,7 @@ public:
     {
         out << "tmp" << m_targetIndex << ": ";
         ESIR::dump(out);
-        out << " const pointer " << std::hex << bitwise_cast<unsigned>(m_value) << std::dec;
+        out << " const pointer " << std::hex << bitwise_cast<unsigned long>(m_value) << std::dec;
     }
 #endif
